Forgot to add receive to router. Most recent test now passes

+++ contracts/core/DEXPair.sol
@@ -68,7 +68,7 @@ contract DEXPair is IDEXPair {
         {
         uint balance0Adjusted = (balance0 * 1000) - (amount0In * 3);
         uint balance1Adjusted = (balance1 * 1000) - (amount1In * 3);
-        require(balance0Adjusted * balance1Adjusted >= (uint(_reserve0 *_reserve1) * 1000**2), 'Error: K');
+        require(balance0Adjusted * balance1Adjusted >= (uint(_reserve0) * uint(_reserve1) * 1000**2), 'Error: K');
         }
 
         _update(balance0, balance1);
