+++ contracts/token/MockWETH.sol
@@ -19,7 +19,7 @@ contract MockWETH is IWETH {
         balanceOf[msg.sender] += msg.value;
         // Deposit(msg.sender, msg.value);
     }
-    function withdraw(uint wad) public {
+    function withdraw(uint256 wad) public {
         require(balanceOf[msg.sender] >= wad);
         balanceOf[msg.sender] -= wad;
         payable(msg.sender).transfer(wad);
